requirements/base.txt
Django==4.1.5
django-celery-beat==2.4.0
django-mptt==0.13.4
django-rest-swagger==2.2.0
django-simple-history==3.0.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.0
django-rest-passwordreset==1.3.0
django-datatables-view==1.20.0
django-dynamic-preferences==1.13.0
django-grappelli==3.0.3
django-cors-headers==3.13.0
django-seed==0.3.1
drf-yasg==1.21.4

psycopg2==2.9.5
sqlparse==0.4.2
typing-extensions==4.0.1
django-filter==21.1
requests==2.27.1
Pillow==9.3.0

python-dateutil==2.8.2
python-dotenv==0.19.2
xhtml2pdf==0.2.8
XlsxWriter==3.0.3
redis==4.3.4
celery==5.2.7
django-redis==3.1.6
reportbro-lib==3.0.5
humanize==4.4.0

six==1.16.0
django-extra-fields==3.0.2
